[refactor] 클린코드 -  함수분리, else예약어 처리, 함수분리리
