+++ index.js
@@ -1,0 +1,4 @@
+import App from "./src/App";
+
+const app = new App();
+app.play();
