+++ src/Play/PlayGame.js
@@ -1,18 +1,20 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
 import randomNum from "./RandomNum.js";
 import isValidInput from "./IsValidInput.js";
+import { count, printHint } from "./Count.js";
 
 //게임 시작 - 랜덤 숫자 생성, 숫자를 입력받기 위한 함수 호출
 const start = async() => {
     MissionUtils.Console.print('숫자 야구 게임을 시작합니다.');
     let answerNumber = randomNum();
-    let isSame = false;
-    while(!isSame){
+    let isPlaying = false;
+    while(!isPlaying){
         let inputNumber = await inputNum();
-        isSame = checkNum(inputNumber, answerNumber);
+        isPlaying = checkNum(inputNumber, answerNumber);
     }
     replay();
 }
+
 
 //사용자에게 숫자 입력받기
 const inputNum = async() => {
@@ -23,45 +25,24 @@ const inputNum = async() => {
     return inputNumber;
 }
 
-//힌트 출력하기
+
+//힌트 계산하기
 const checkNum = (inputNumber, answerNumber) => {
     let userNumber = inputNumber.toString().split('').map(Number);
 
-    let strike = 0;
-    let ball = 0;
+    let hintcount = {
+        strike : 0,
+        ball : 0
+    };
 
     // 정답과 사용자 입력값 비교
     for(let i = 0; i < answerNumber.length; i++){
         for(let j = 0; j < userNumber.length; j++){
-            if(answerNumber[i] === userNumber[j]){
-                if(i === j){
-                    strike++;
-                } else {
-                    ball++;
-                }
-            }
+            count(userNumber, answerNumber, i, j, hintcount);
         }
     }
 
-    let hint = [];
-    if(ball !== 0) {
-        hint.push(`${ball}볼`);
-    }
-    if (strike !== 0){
-        hint.push(`${strike}스트라이크`);
-    }
-    if(ball === 0 && strike === 0) {
-        hint.push("낫싱");
-    }
-    
-    MissionUtils.Console.print(hint.join(' '));
-
-    if(strike === 3){
-        MissionUtils.Console.print("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
-        return true;
-    } else {
-        return false;
-    }
+    return printHint(hintcount);
 }
 
 
@@ -69,9 +50,10 @@ const checkNum = (inputNumber, answerNumber) => {
 const replay = async() => {
     const replay = await MissionUtils.Console.readLineAsync("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
     if (replay === '1') {
-        start();
+        return start();
     } else if (replay === '2') {
         //gmae end
+        return false;
     } else {
         throw new Error("[ERROR] 올바른 숫자를 입력해주세요.");
     }
